Move composer-scaffold to init-site shared, add circle ci CMS_PLATFORM check
