+++ lib/init-site/d8/init-site-d8.js
@@ -103,7 +103,7 @@ function InitSiteD8(args) {
     // Add required default drupal modules and themes.
     log(chalk.yellow("Try to run Composer commands\n"), 0);
 
-    let Composer = require('../../comp-install-scaffold/comp-install-scaffold.js');
+    let Composer = require('../shared/composer-install-scaffold.js');
     let composer = new Composer({ cms: "D8", packages: ['adminTheme', 'devModules', 'baseBluecadet', 'baseWeb'] });
     composer.run();
