+++ lib/init-site/init-site.js
@@ -1,11 +1,11 @@
 let log = require('../utils/logger.js').getInstance().log;
 const creds = require('../utils/credentials.js');
+const setCircleCMS = require('./shared/circle-ci-setup.js');
 
 function InitSite(args) {
   this.args = args;
 
   this.run = () => {
-    // console.log('TODO: init-site');
     // console.log(args);
 
     // github credentials (global)
@@ -25,12 +25,16 @@ function InitSite(args) {
           case 'Drupal 8: Default':
             let d8InitSiteInstaller = require('./d8/init-site-d8.js');
             new d8InitSiteInstaller(this.args).initSiteSetup();
+            setCircleCMS('D8');
             break;
 
           case 'WordPress: Circle CI':
           case 'WordPress: Default':
             let wpInit = require('./wp/init-site-wp.js');
-            wpInit();
+            setCircleCMS('WP')
+              .then(() => {
+                wpInit();
+              });
             break;
 
           default:
