+++ lib/init-site/shared/composer-install-scaffold.js
@@ -1,11 +1,8 @@
-const os = require('os');
-const path = require('path');
 const fsx = require('fs-extra');
-let log = require('../utils/logger.js').getInstance().log;
+const log = require('../../utils/logger.js').getInstance().log;
 const chalk = require('chalk');
-const creds = require('../utils/credentials.js');
 const spawnSync = require('child_process').spawnSync;
-const compPackages = require('../utils/packages/packages.js').PACKAGES;
+const compPackages = require('../../utils/packages/packages.js').PACKAGES;
 
 function Composer(args) {
   this.args = args;
