+++ lib/init-site/wp/init-site-wp.js
@@ -49,9 +49,7 @@ module.exports = function(args) {
 
 
   // Add required default drupal modules and themes.
-  log(chalk.blue("...running Composer commands...\n"), 0);
-
-  let Composer = require('../../comp-install-scaffold/comp-install-scaffold.js');
+  let Composer = require('../shared/composer-install-scaffold.js');
   let composer = new Composer({ cms: "WP", packages: ['web'] });
   composer.run();
