+++ package.json
@@ -41,6 +41,7 @@
     "request": "^2.88.0",
     "rsync": "^0.6.1",
     "semver": "^6.3.0",
+    "yaml": "^1.7.2",
     "yargs": "^13.2.4",
     "zlib": "^1.0.5"
   }
